refactor : folderList를 불러오는 함수를 folderPage로 이동
