+++ src/components/card/CardList.js
@@ -5,9 +5,11 @@ import './CardList.css';
 function CardList({ links }) {
   return (
     <div className="card-container">
-      {links.map((link) => (
-        <Card key={link.id} data={link} />
-      ))}
+      {links.length > 0 ? (
+        links.map((link) => <Card key={link.id} data={link} />)
+      ) : (
+        <span className="no-link">저장된 링크가 없습니다.</span>
+      )}
     </div>
   );
 }
